tests/wpt/web-platform-tests/pointerevents/pointerevent_pointerup-manual.html

<!doctype html>
<html>
    <head>
        <title>pointerup</title>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
    </head>
    <body onload="run()">
        <h1>pointerup</h1>
        <h4>Test Description: This test checks if pointerup event triggers. Press mouse left button and release it over the black rectangle or tap it if you are using a touchscreen.</h4>
        <div id="target0" style="background:black"></div>
        <script>
            var detected_pointertypes = {};
            var test_pointerEvent = async_test("pointerup event received");
            // showPointerTypes is defined in pointerevent_support.js
            // Requirements: the callback function will reference the test_pointerEvent object and
            // will fail unless the async_test is created with the var name "test_pointerEvent".
            add_completion_callback(showPointerTypes);

            function run() {
                var target0 = document.getElementById("target0");

                on_event(target0, "pointerup", function (event) {
                    detected_pointertypes[event.pointerType] = true;
                    test_pointerEvent.step(function () {
                        check_PointerEvent(event);
                        assert_equals(event.type, "pointerup", "Pointer Event received.");
                    });
                    test_pointerEvent.done();
                });
            }
        </script>
        <h1>Pointer Events pointerup tests</h1>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Refresh the page to run the tests again with a different pointer type.</p>
        </div>
        <div id="log"></div>
    </body>
</html>
